adjusted the name and id to match the backend for the signUp form.
finalized the authentication middleware. Tested to make sure that it works within various routes. Can ensure that a user is properly authenticated based on the session in the database.

finalized the authorization middleware. Tested to make sure that it works withing various routes. Can ensure that a user is properly authorized and routed to the appropriate dashboard.

cleaned up the userRoutes, and indexRoutes so that users are taken to the proper pages as expected.

Created various test files that can be used for myself to test routing and